+++ lib/js-interpreter/interpreter.d.ts
@@ -1,6 +1,0 @@
-declare export class Interpreter {
-  value: any;
-  constructor(code: string);
-  appendCode(code: string): void;
-  run(): void;
-}